+++ pkg/scalers/redis_streams_scaler.go
@@ -12,6 +12,7 @@ import (
 	v2 "k8s.io/api/autoscaling/v2"
 	"k8s.io/metrics/pkg/apis/external_metrics"
 
+	kedav1alpha1 "github.com/kedacore/keda/v2/apis/keda/v1alpha1"
 	"github.com/kedacore/keda/v2/pkg/scalers/scalersconfig"
 	kedautil "github.com/kedacore/keda/v2/pkg/util"
 )
@@ -59,11 +60,16 @@ type redisStreamsMetadata struct {
 	ActivationLagCount        int64               `keda:"name=activationLagCount,  order=triggerMetadata, default=0"`
 	MetadataEnableTLS         string              `keda:"name=enableTLS,           order=triggerMetadata, optional"`
 	AuthParamEnableTLS        string              `keda:"name=tls,                 order=authParams, optional"`
+	podIdentity               kedav1alpha1.AuthPodIdentity
 }
 
 func (r *redisStreamsMetadata) Validate() error {
 	err := validateRedisAddress(&r.ConnectionInfo)
 	if err != nil {
+		return err
+	}
+
+	if err := validateRedisEntraIDCompatibility(r.ConnectionInfo, r.podIdentity); err != nil {
 		return err
 	}
 
@@ -116,7 +122,7 @@ func NewRedisStreamsScaler(ctx context.Context, isClustered, isSentinel bool, co
 }
 
 func createClusteredRedisStreamsScaler(ctx context.Context, meta *redisStreamsMetadata, metricType v2.MetricTargetType, logger logr.Logger) (Scaler, error) {
-	client, err := getRedisClusterClient(ctx, meta.ConnectionInfo)
+	client, err := getRedisClusterClient(ctx, meta.ConnectionInfo, meta.podIdentity, logger)
 
 	if err != nil {
 		return nil, fmt.Errorf("connection to redis cluster failed: %w", err)
@@ -151,7 +157,7 @@ func createSentinelRedisStreamsScaler(ctx context.Context, meta *redisStreamsMet
 }
 
 func createRedisStreamsScaler(ctx context.Context, meta *redisStreamsMetadata, metricType v2.MetricTargetType, logger logr.Logger) (Scaler, error) {
-	client, err := getRedisClient(ctx, meta.ConnectionInfo, meta.DatabaseIndex)
+	client, err := getRedisClient(ctx, meta.ConnectionInfo, meta.DatabaseIndex, meta.podIdentity, logger)
 	if err != nil {
 		return nil, fmt.Errorf("connection to redis failed: %w", err)
 	}
@@ -284,6 +290,7 @@ var (
 func parseRedisStreamsMetadata(config *scalersconfig.ScalerConfig) (*redisStreamsMetadata, error) {
 	meta := &redisStreamsMetadata{}
 	meta.triggerIndex = config.TriggerIndex
+	meta.podIdentity = config.PodIdentity
 	if err := config.TypedConfig(meta); err != nil {
 		return nil, fmt.Errorf("error parsing redis stream metadata: %w", err)
 	}